🚧 WIP: Extract owid-etl framework library
This PR begins extracting core ETL framework functionality into a standalone
`owid-etl` package in `lib/etl/`. The goal is to provide reusable ETL
infrastructure that can be used independently of OWID-specific code.

## What's been done

### New `lib/etl/` package structure
- `owid/etl/config.py` - ETLConfig dataclass for user configuration
- `owid/etl/dag.py` - DAG loading and graph operations
- `owid/etl/steps.py` - Step protocol and implementations (DataStep, SnapshotStep, etc.)
- `owid/etl/execution.py` - Step execution engine
- `owid/etl/snapshot.py` - Snapshot class for data ingestion
- `owid/etl/cli.py` - Basic CLI runner

### Step Registry Pattern
Introduced a registry pattern for step types:
```python
@register_step("data")
class DataStep: ...

@register_step("grapher")
class GrapherStep: ...  # Defined in etl/steps/__init__.py (OWID-specific)
```

This allows external code to register custom step types.

### Re-export wrappers
- `etl/steps/__init__.py` - Re-exports from owid.etl.steps + defines GrapherStep
- `etl/snapshot.py` - Re-exports from owid.etl.snapshot (WIP - has circular import)
- `etl/dag_helpers.py` - Re-exports Graph type

## Known Issues / TODO

1. **Circular import**: `lib/etl/owid/etl/steps.py` imports from `etl.helpers`
   which imports from `etl.snapshot`, causing a circular dependency when
   importing snapshot.

2. **Type checking**: pyright errors need resolution (extraPaths configured
   but some Protocol attributes missing)

3. **OWID dependencies**: The lib still depends on `etl.config`, `etl.paths`,
   `etl.files` - these need to be made configurable or moved to lib

4. **Tests**: Only basic DAG and config tests exist, need more coverage

## Example usage (target API)
```python
from owid.etl import ETLConfig, run_dag, set_config

config = ETLConfig(
    base_dir=Path("./"),
    steps_dir=Path("./steps/data"),
    dag_file=Path("./dag.yml"),
)
set_config(config)

run_dag(steps=["data://garden/example/2024/dataset"])
```
