+++ examples/cherry_blossom/data/garden/example/2024-01-01/cherry_blossom/cherry_blossom.meta.json
@@ -1,0 +1,33 @@
+{
+  "short_name": "cherry_blossom",
+  "dataset": {
+    "channel": "garden",
+    "namespace": "example",
+    "short_name": "cherry_blossom",
+    "title": "Cherry Blossom Bloom Dates (Processed)",
+    "is_public": true,
+    "version": "2024-01-01",
+    "non_redistributable": false,
+    "source_checksum": "7100a1007be6b9d13fd8addd34652fc2"
+  },
+  "primary_key": [
+    "country",
+    "year"
+  ],
+  "dimensions": [
+    {
+      "name": "country",
+      "slug": "country"
+    },
+    {
+      "name": "year",
+      "slug": "year"
+    }
+  ],
+  "fields": {
+    "country": {},
+    "year": {},
+    "bloom_date": {},
+    "bloom_day_of_year": {}
+  }
+}
