+++ examples/cherry_blossom/data/meadow/example/2024-01-01/cherry_blossom/cherry_blossom.meta.json
@@ -1,0 +1,32 @@
+{
+  "short_name": "cherry_blossom",
+  "dataset": {
+    "channel": "meadow",
+    "namespace": "example",
+    "short_name": "cherry_blossom",
+    "title": "Cherry Blossom Bloom Dates",
+    "is_public": true,
+    "version": "2024-01-01",
+    "non_redistributable": false,
+    "source_checksum": "af0104dbd463bff70ca3ad099977dd28"
+  },
+  "primary_key": [
+    "country",
+    "year"
+  ],
+  "dimensions": [
+    {
+      "name": "country",
+      "slug": "country"
+    },
+    {
+      "name": "year",
+      "slug": "year"
+    }
+  ],
+  "fields": {
+    "country": {},
+    "year": {},
+    "bloom_date": {}
+  }
+}
