+++ lib/etl/Makefile
@@ -1,0 +1,8 @@
+#
+#  Makefile
+#
+
+
+include ../../default.mk
+
+SRC = owid tests
