+++ lib/etl/pyproject.toml
@@ -1,0 +1,48 @@
+[project]
+name = "owid-etl"
+version = "0.1.0"
+description = "Core ETL framework for building data pipelines."
+readme = "README.md"
+authors = [
+    {name = "Our World in Data", email = "tech@ourworldindata.org"},
+]
+license = "MIT"
+requires-python = ">=3.10, <4.0"
+dependencies = [
+    "click>=8.0",
+    "PyYAML>=6.0",
+    "structlog>=21.0",
+    "owid-catalog",
+]
+
+[tool.uv]
+dev-dependencies = [
+    "pytest>=6.2.5",
+    "pytest-cov>=2.12.1",
+    "watchdog>=2.1.5",
+    "pyright==1.1.373",
+    "ruff==0.8.6",
+]
+
+[tool.uv.sources]
+owid-catalog = { path = "../catalog", editable = true }
+
+[tool.ruff]
+extend = "../../pyproject.toml"
+
+[tool.pyright]
+reportIncompatibleMethodOverride = "none"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["owid"]
+
+[project.urls]
+Homepage = "https://github.com/owid/etl/tree/master/lib/etl"
+Repository = "https://github.com/owid/etl.git"
+
+[project.scripts]
+owid-etl = "owid.etl.cli:main"
